Moving extension into a separate file
